Fix for 32-bit build.

+++ LLVMTypeUtils.cpp
@@ -163,7 +163,7 @@ void createCollectionCopy(const TypeRef& collection_type, llvm::Value* dest_ptr,
 		const unsigned int type_alignment = params.target_data->getABITypeAlignment(collection_type->LLVMType(*params.module));
 
 		llvm::Type* llvm_type = collection_type->LLVMType(*params.module);
-		const size_t size_B = params.target_data->getTypeAllocSize(llvm_type);
+		const uint64_t size_B = params.target_data->getTypeAllocSize(llvm_type);
 		llvm::Value* size = llvm::ConstantInt::get(*params.context, llvm::APInt(64, size_B, /*signed=*/false));
 		params.builder->CreateMemCpy(dest_ptr, src_ptr, size,
 			type_alignment // align
